+++ src/+bl12014/+ui/HeightSensor.m
@@ -44,7 +44,7 @@ classdef HeightSensor < mic.Base
         dWidthName = 70
         dWidthUnit = 80
         dWidthVal = 75
-        dWidthPadUnit = 280
+        dWidthPadUnit = 277
         
         configStageY
         configMeasPointVolts